third_party/blink/web_tests/external/wpt/custom-state-pseudo-class/state-css-selector.tentative-expected.txt

This is a testharness.js-based test.
[PASS] state selector has no influence when state is not applied
[PASS] state selector has no influence on sibling selectors when not applied
[PASS] state selector has influence when state is applied
[PASS] state selector influences siblings when state is applied
[FAIL] state selector influences has() when state is applied
  assert_equals: expected "rgb(0, 0, 255)" but got "rgb(0, 0, 0)"
[PASS] state selector only applies on given ident
[PASS] state selector only applies to siblings on given ident
[PASS] state selector only applies to has() on given ident
Harness: the test ran to completion.
